third_party/blink/web_tests/html5lib/generated/run-tests18-write-expected.txt

CONSOLE ERROR: Uncaught SyntaxError: Unexpected token '<'
CONSOLE ERROR: Uncaught SyntaxError: Unexpected token '<'
CONSOLE ERROR: Uncaught SyntaxError: Unexpected token '<'
CONSOLE ERROR: Uncaught SyntaxError: Unexpected token '<'
../resources/tests18.dat: PASS
